Move base token tests in a separate file
